UI: Stop using Redux to store view state
Fixes https://github.com/product-os/jellyfish/issues/7931

This change introduces `LiveCollection`, a component that wraps
together lens selection, filtering, sorting, pagination and live
updates.
Result state is stored inside this component and passed down to the
chosen lens renderer, instead of being stored globally in redux.
Seperating out this functionality as a seperate component also makes it
easier to re-use it elsewhere in the future (e.g. link tabs or inbox).

Change-type: patch
