+++ apps/ui/lib/lens/common/LiveCollection/Header/SliceOptions.tsx
